[Feature #160425661] move the RedirectCount... component to its own file
