+++ src/components/SignUp/index.js
@@ -1,36 +1,13 @@
 import React from 'react';
 import { connect } from 'react-redux';
-import { Link } from 'react-router-dom';
 import PropTypes from 'prop-types';
 import NavBar from '../commons/NavBar';
 import Main from '../commons/Main/index';
+import handleBasicSignUp, { signUpResetState } from '../../actions/signUpActions';
+import RedirectCountdownAlert from './RedirectCountdownAlert';
+import SignUpFormCard from './SignUpFormCard';
 import Footer from '../commons/Footer/index';
 import './SignUp.css';
-import SignUpFormCard from './SignUpFormCard';
-import handleBasicSignUp, { signUpResetState } from '../../actions/signUpActions';
-
-/** *This is the alert that will show the user that they will be redirected after a few seconds
- * @param secondsToRedirect: {number} The seconds remaining before redirecting */
-const RedirectCountdownAlert = ({ secondsToRedirect }) => (
-  <div
-    className="sign-up-success-message alert alert-warning text-center"
-    style={{ marginTop: '5em' }}
-  >
-      We will redirect you to the
-    {' '}
-    <Link to="/login">login</Link>
-    {' '}
-      page in
-    {' '}
-    <strong>{secondsToRedirect}</strong>
-    {' '}
-      seconds
-  </div>
-);
-
-RedirectCountdownAlert.propTypes = {
-  secondsToRedirect: PropTypes.number.isRequired,
-};
 
 /** *This is the class that represents the signup page and its state */
 class SignUp extends React.Component {
